requirements.txt
streamlit>=1.30.0
PyPDF2>=3.0.0
python-docx>=1.0.0
google-genai>=0.3.0
python-dotenv>=1.0.0